requirements.txt
pyrogram>=2.0.0
python-dotenv>=1.0.0
tgcrypto>=1.2.5
aiohttp>=3.9.0
